pyproject.toml
[tool.poetry]
name = "small-fileread-service"
version = "0.1.0"
description = ""
authors = ["Izbushko <izbaizbushka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
python-multipart = "^0.0.9"
gunicorn = "^21.2.0"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.2.1"
python-docx = "^1.1.0"
jinja2 = "^3.1.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
